requirements.txt
# VisionPreprocessor dependencies
opencv-python
numpy
Pillow
pdf2image

# VisionService dependencies
torch
transformers
scikit-image
torchvision

# NLPModule dependencies
pytesseract
PyPDF2
pandas
nltk
fastapi
uvicorn
python-multipart
pandas
scikit-learn